+++ pandapower/std_types.py
@@ -107,7 +107,6 @@ def create_std_types(net, data, element="line", overwrite=True, check_required=T
 
     INPUT:
         **net** - The pandapower network
-
         **data** - dictionary of standard type parameter sets
 
         **element** - "line", "trafo" or "trafo3w"
@@ -224,7 +223,7 @@ def available_std_types(net, element="line"):
             return std_types.convert_objects()
 
 
-def parameter_from_std_type(net, parameter, element="line", fill=None):
+def parameter_from_std_type(net, parameter, element="line",fill=None):
     """
     Loads standard types data for a parameter, which can be used to add an additional parameter,
     that is not included in the original pandapower datastructure but is available in the standard
@@ -301,8 +300,8 @@ def find_std_type_by_parameter(net, data, element="line", epsilon=0.):
     OUTPUT:
         **fitting_types** - list of fitting types or empty list
     """
-    assert epsilon >= 0
     fitting_types = []
+    assert epsilon >= 0
     for name, stp in net.std_types[element].items():
         for p, v in list(data.items()):
             if isinstance(v, float):
@@ -314,6 +313,43 @@ def find_std_type_by_parameter(net, data, element="line", epsilon=0.):
             fitting_types.append(name)
     return fitting_types
 
+def find_std_type_alternative(net, data, element = "line", voltage_rating = "", epsilon = 0.):
+    """
+        Searches for a std_type that fits all values given in the standard types library with the margin of
+        epsilon.
+
+        INPUT:
+            **net** - pandapower network
+
+            **data** - dictionary of standard type parameters
+
+            **element** - type of element ("line" or "trafo")
+
+            **voltage_rating** - voltage rating of the cable ("HV" or "MV" or "LV")
+
+            **epsilon** - tolerance margin for parameter comparison
+
+        OUTPUT:
+            **fitting_types** - list of fitting types or empty list
+        """
+
+    assert epsilon >= 0
+    linetypes = basic_line_std_types()
+    possible_alternatives = []
+    fitting_types = []
+    for p, v in linetypes.items():
+        if voltage_rating == v.get("voltage_rating"):
+            possible_alternatives.append((p, v))
+    for name, stp in possible_alternatives:
+        for p, v in list(data.items()):
+            if isinstance(v, float):
+                if abs(v - stp[p]) > epsilon:
+                    break
+            elif stp[p] != v:
+                break
+        else:
+            fitting_types.append(name)
+    return fitting_types
 
 def add_zero_impedance_parameters(net):
     """
@@ -380,7 +416,8 @@ def basic_line_std_types():
             "max_i_ka": 0.142,
             "type": "cs",
             "q_mm2": 50,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+            "voltage_rating": "LV"},
         "NAYY 4x120 SE":
         {"c_nf_per_km": 264,
             "r_ohm_per_km": 0.225,
@@ -388,7 +425,8 @@ def basic_line_std_types():
             "max_i_ka": 0.242,
             "type": "cs",
             "q_mm2": 120,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "LV"},
         "NAYY 4x150 SE":
         {"c_nf_per_km": 261,
             "r_ohm_per_km": 0.208,
@@ -396,7 +434,8 @@ def basic_line_std_types():
             "max_i_ka": 0.270,
             "type": "cs",
             "q_mm2": 150,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "LV"},
 
         # Medium Voltage
         "NA2XS2Y 1x95 RM/25 12/20 kV":
@@ -406,7 +445,8 @@ def basic_line_std_types():
             "max_i_ka": 0.252,
             "type": "cs",
             "q_mm2": 95,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "NA2XS2Y 1x185 RM/25 12/20 kV":
         {"c_nf_per_km": 273,
             "r_ohm_per_km": 0.161,
@@ -414,7 +454,8 @@ def basic_line_std_types():
             "max_i_ka": 0.362,
             "type": "cs",
             "q_mm2": 185,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "NA2XS2Y 1x240 RM/25 12/20 kV":
         {"c_nf_per_km": 304,
             "r_ohm_per_km": 0.122,
@@ -422,7 +463,8 @@ def basic_line_std_types():
             "max_i_ka": 0.421,
             "type": "cs",
             "q_mm2": 240,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "NA2XS2Y 1x95 RM/25 6/10 kV":
         {"c_nf_per_km": 315,
             "r_ohm_per_km": 0.313,
@@ -430,7 +472,8 @@ def basic_line_std_types():
             "max_i_ka": 0.249,
             "type": "cs",
             "q_mm2": 95,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "NA2XS2Y 1x185 RM/25 6/10 kV":
         {"c_nf_per_km": 406,
             "r_ohm_per_km": 0.161,
@@ -438,7 +481,8 @@ def basic_line_std_types():
             "max_i_ka": 0.358,
             "type": "cs",
             "q_mm2": 185,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "NA2XS2Y 1x240 RM/25 6/10 kV":
         {"c_nf_per_km": 456,
             "r_ohm_per_km": 0.122,
@@ -446,7 +490,8 @@ def basic_line_std_types():
             "max_i_ka": 0.416,
             "type": "cs",
             "q_mm2": 240,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         # additional MV cables
         "NA2XS2Y 1x150 RM/25 12/20 kV":
         {"c_nf_per_km": 250,
@@ -455,7 +500,8 @@ def basic_line_std_types():
             "max_i_ka": 0.319,
             "type": "cs",
             "q_mm2": 150,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "NA2XS2Y 1x120 RM/25 12/20 kV":
         {"c_nf_per_km": 230,
             "r_ohm_per_km": 0.253,
@@ -463,7 +509,8 @@ def basic_line_std_types():
             "max_i_ka": 0.283,
             "type": "cs",
             "q_mm2": 120,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "NA2XS2Y 1x70 RM/25 12/20 kV":
         {"c_nf_per_km": 190,
             "r_ohm_per_km": 0.443,
@@ -471,7 +518,8 @@ def basic_line_std_types():
             "max_i_ka": 0.220,
             "type": "cs",
             "q_mm2": 70,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "NA2XS2Y 1x150 RM/25 6/10 kV":
         {"c_nf_per_km": 360,
             "r_ohm_per_km": 0.206,
@@ -479,7 +527,8 @@ def basic_line_std_types():
             "max_i_ka": 0.315,
             "type": "cs",
             "q_mm2": 150,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "NA2XS2Y 1x120 RM/25 6/10 kV":
         {"c_nf_per_km": 340,
             "r_ohm_per_km": 0.253,
@@ -487,7 +536,8 @@ def basic_line_std_types():
             "max_i_ka": 0.280,
             "type": "cs",
             "q_mm2": 120,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "NA2XS2Y 1x70 RM/25 6/10 kV":
         {"c_nf_per_km": 280,
             "r_ohm_per_km": 0.443,
@@ -495,7 +545,8 @@ def basic_line_std_types():
             "max_i_ka": 0.217,
             "type": "cs",
             "q_mm2": 70,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
 
         # High Voltage
         "N2XS(FL)2Y 1x120 RM/35 64/110 kV":
@@ -505,7 +556,8 @@ def basic_line_std_types():
             "max_i_ka": 0.366,
             "type": "cs",
             "q_mm2": 120,
-            "alpha": alpha_cu},
+            "alpha": alpha_cu,
+         "voltage_rating": "HV"},
         "N2XS(FL)2Y 1x185 RM/35 64/110 kV":
         {"c_nf_per_km": 125,
             "r_ohm_per_km": 0.099,
@@ -513,7 +565,8 @@ def basic_line_std_types():
             "max_i_ka": 0.457,
             "type": "cs",
             "q_mm2": 185,
-            "alpha": alpha_cu},
+            "alpha": alpha_cu,
+         "voltage_rating": "HV"},
         "N2XS(FL)2Y 1x240 RM/35 64/110 kV":
         {"c_nf_per_km": 135,
             "r_ohm_per_km": 0.075,
@@ -521,7 +574,8 @@ def basic_line_std_types():
             "max_i_ka": 0.526,
             "type": "cs",
             "q_mm2": 240,
-            "alpha": alpha_cu},
+            "alpha": alpha_cu,
+         "voltage_rating": "HV"},
         "N2XS(FL)2Y 1x300 RM/35 64/110 kV":
         {"c_nf_per_km": 144,
             "r_ohm_per_km": 0.060,
@@ -529,7 +583,8 @@ def basic_line_std_types():
             "max_i_ka": 0.588,
             "type": "cs",
             "q_mm2": 300,
-            "alpha": alpha_cu},
+            "alpha": alpha_cu,
+         "voltage_rating": "HV"},
 
         # Overhead Lines, all from S.742f, Heuck: Elektrische Energieversorgung -
         # Vierweg+Teubner 2013
@@ -543,7 +598,8 @@ def basic_line_std_types():
             "max_i_ka": 0.105,
             "type": "ol",
             "q_mm2": 16,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "LV"},
         "24-AL1/4-ST1A 0.4":
         {"c_nf_per_km": 11.25,
             "r_ohm_per_km": 1.2012,
@@ -551,7 +607,8 @@ def basic_line_std_types():
             "max_i_ka": 0.140,
             "type": "ol",
             "q_mm2": 24,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "LV"},
         "48-AL1/8-ST1A 0.4":
         {"c_nf_per_km": 12.2,
             "r_ohm_per_km": 0.5939,
@@ -559,7 +616,8 @@ def basic_line_std_types():
             "max_i_ka": .210,
             "type": "ol",
             "q_mm2": 48,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "LV"},
         "94-AL1/15-ST1A 0.4":
         {"c_nf_per_km": 13.2,
             "r_ohm_per_km": 0.3060,
@@ -567,7 +625,8 @@ def basic_line_std_types():
             "max_i_ka": 0.350,
             "type": "ol",
             "q_mm2": 94,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "LV"},
 
         # Medium Voltage
         "34-AL1/6-ST1A 10.0":
@@ -577,7 +636,8 @@ def basic_line_std_types():
             "max_i_ka": 0.170,
             "type": "ol",
             "q_mm2": 34,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "48-AL1/8-ST1A 10.0":
         {"c_nf_per_km": 10.1,
             "r_ohm_per_km": 0.5939,
@@ -585,7 +645,8 @@ def basic_line_std_types():
             "max_i_ka": 0.210,
             "type": "ol",
             "q_mm2": 48,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "70-AL1/11-ST1A 10.0":
         {"c_nf_per_km": 10.4,
             "r_ohm_per_km": 0.4132,
@@ -593,7 +654,8 @@ def basic_line_std_types():
             "max_i_ka": 0.290,
             "type": "ol",
             "q_mm2": 70,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "94-AL1/15-ST1A 10.0":
         {"c_nf_per_km": 10.75,
             "r_ohm_per_km": 0.3060,
@@ -601,7 +663,8 @@ def basic_line_std_types():
             "max_i_ka": 0.350,
             "type": "ol",
             "q_mm2": 94,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "122-AL1/20-ST1A 10.0":
         {"c_nf_per_km": 11.1,
             "r_ohm_per_km": 0.2376,
@@ -609,7 +672,8 @@ def basic_line_std_types():
             "max_i_ka": 0.410,
             "type": "ol",
             "q_mm2": 122,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "149-AL1/24-ST1A 10.0":
         {"c_nf_per_km": 11.25,
             "r_ohm_per_km": 0.1940,
@@ -617,7 +681,8 @@ def basic_line_std_types():
             "max_i_ka": 0.470,
             "type": "ol",
             "q_mm2": 149,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "34-AL1/6-ST1A 20.0":
         {"c_nf_per_km": 9.15,
             "r_ohm_per_km": 0.8342,
@@ -625,7 +690,8 @@ def basic_line_std_types():
             "max_i_ka": 0.170,
             "type": "ol",
             "q_mm2": 34,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "48-AL1/8-ST1A 20.0":
         {"c_nf_per_km": 9.5,
             "r_ohm_per_km": 0.5939,
@@ -633,7 +699,8 @@ def basic_line_std_types():
             "max_i_ka": 0.210,
             "type": "ol",
             "q_mm2": 48,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "70-AL1/11-ST1A 20.0":
         {"c_nf_per_km": 9.7,
             "r_ohm_per_km": 0.4132,
@@ -641,7 +708,8 @@ def basic_line_std_types():
             "max_i_ka": 0.290,
             "type": "ol",
             "q_mm2": 70,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "94-AL1/15-ST1A 20.0":
         {"c_nf_per_km": 10,
             "r_ohm_per_km": 0.3060,
@@ -649,7 +717,8 @@ def basic_line_std_types():
             "max_i_ka": 0.350,
             "type": "ol",
             "q_mm2": 94,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "122-AL1/20-ST1A 20.0":
         {"c_nf_per_km": 10.3,
             "r_ohm_per_km": 0.2376,
@@ -657,7 +726,8 @@ def basic_line_std_types():
             "max_i_ka": 0.410,
             "type": "ol",
             "q_mm2": 122,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "149-AL1/24-ST1A 20.0":
         {"c_nf_per_km": 10.5,
             "r_ohm_per_km": 0.1940,
@@ -665,7 +735,8 @@ def basic_line_std_types():
             "max_i_ka": 0.470,
             "type": "ol",
             "q_mm2": 149,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "184-AL1/30-ST1A 20.0":
         {"c_nf_per_km": 10.75,
             "r_ohm_per_km": 0.1571,
@@ -673,7 +744,8 @@ def basic_line_std_types():
             "max_i_ka": 0.535,
             "type": "ol",
             "q_mm2": 184,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
         "243-AL1/39-ST1A 20.0":
         {"c_nf_per_km": 11,
             "r_ohm_per_km": 0.1188,
@@ -681,7 +753,8 @@ def basic_line_std_types():
             "max_i_ka": 0.645,
             "type": "ol",
             "q_mm2": 243,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "MV"},
 
         # High Voltage
         # c acd x values are estimated for 4 m conductor distance, single bundle and "Donaumast"
@@ -692,7 +765,8 @@ def basic_line_std_types():
             "max_i_ka": 0.210,
             "type": "ol",
             "q_mm2": 48,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "HV"},
         "70-AL1/11-ST1A 110.0":
         {"c_nf_per_km": 8.4,
             "r_ohm_per_km": 0.4132,
@@ -700,7 +774,8 @@ def basic_line_std_types():
             "max_i_ka": 0.290,
             "type": "ol",
             "q_mm2": 70,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "HV"},
         "94-AL1/15-ST1A 110.0":
         {"c_nf_per_km": 8.65,
             "r_ohm_per_km": 0.3060,
@@ -708,7 +783,8 @@ def basic_line_std_types():
             "max_i_ka": 0.350,
             "type": "ol",
             "q_mm2": 94,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "HV"},
         "122-AL1/20-ST1A 110.0":
         {"c_nf_per_km": 8.5,
             "r_ohm_per_km": 0.2376,
@@ -716,7 +792,8 @@ def basic_line_std_types():
             "max_i_ka": 0.410,
             "type": "ol",
             "q_mm2": 122,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "HV"},
         "149-AL1/24-ST1A 110.0":
         {"c_nf_per_km": 8.75,
             "r_ohm_per_km": 0.1940,
@@ -724,7 +801,8 @@ def basic_line_std_types():
             "max_i_ka": 0.470,
             "type": "ol",
             "q_mm2": 149,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "HV"},
         "184-AL1/30-ST1A 110.0":
         {"c_nf_per_km": 8.8,
             "r_ohm_per_km": 0.1571,
@@ -732,7 +810,8 @@ def basic_line_std_types():
             "max_i_ka": 0.535,
             "type": "ol",
             "q_mm2": 184,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "HV"},
         "243-AL1/39-ST1A 110.0":
         {"c_nf_per_km": 9,
             "r_ohm_per_km": 0.1188,
@@ -740,7 +819,8 @@ def basic_line_std_types():
             "max_i_ka": 0.645,
             "type": "ol",
             "q_mm2": 243,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "HV"},
         "305-AL1/39-ST1A 110.0":
         {"c_nf_per_km": 9.2,
             "r_ohm_per_km": 0.0949,
@@ -748,7 +828,8 @@ def basic_line_std_types():
             "max_i_ka": 0.74,
             "type": "ol",
             "q_mm2": 305,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "HV"},
         "490-AL1/64-ST1A 110.0":
         {"c_nf_per_km": 9.75,
             "r_ohm_per_km": 0.059,
@@ -756,7 +837,8 @@ def basic_line_std_types():
             "max_i_ka": 0.960,
             "type": "ol",
             "q_mm2": 490,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "HV"},
         "679-AL1/86-ST1A 110.0":
         {"c_nf_per_km": 9.95,
             "r_ohm_per_km": 0.042,
@@ -764,7 +846,8 @@ def basic_line_std_types():
             "max_i_ka": 1.150,
             "type": "ol",
             "q_mm2": 679,
-            "alpha": alpha_al},
+            "alpha": alpha_al,
+         "voltage_rating": "HV"},
 
         # Transmission System
         # The following values of c and x depend on the geometries of the  overhead line
@@ -777,7 +860,8 @@ def basic_line_std_types():
          "max_i_ka": 0.96,
          "type": "ol",
          "q_mm2": 490,
-         "alpha": alpha_al},
+         "alpha": alpha_al,
+         "voltage_rating": "HV"},
         "679-AL1/86-ST1A 220.0":
         {"c_nf_per_km": 11.7,
          "r_ohm_per_km": 0.042,
@@ -785,7 +869,8 @@ def basic_line_std_types():
          "max_i_ka": 1.150,
          "type": "ol",
          "q_mm2": 679,
-         "alpha": alpha_al},
+         "alpha": alpha_al,
+         "voltage_rating": "HV"},
         "490-AL1/64-ST1A 380.0":
         {"c_nf_per_km": 11,
          "r_ohm_per_km": 0.059,
@@ -793,7 +878,8 @@ def basic_line_std_types():
          "max_i_ka": 0.96,
          "type": "ol",
          "q_mm2": 490,
-         "alpha": alpha_al},
+         "alpha": alpha_al,
+         "voltage_rating": "HV"},
         "679-AL1/86-ST1A 380.0":
         {"c_nf_per_km": 14.6,
          "r_ohm_per_km": 0.042,
@@ -801,7 +887,8 @@ def basic_line_std_types():
          "max_i_ka": 1.150,
          "type": "ol",
          "q_mm2": 679,
-         "alpha": alpha_al}
+         "alpha": alpha_al,
+         "voltage_rating": "HV"}
     }
     return linetypes